Health Script
Added a script that is funcionable for players and enemies that destroys the object it's attached to once it's damaged enough.
It is also set up for invulnerability after being hit, and a health cap.

Aswell, I added a starter enemy triangle called enemy_shooter that currently has no logic and a Tag "Enemy"