Point And Shoot Script
Added a point and shoot scrit addaple to enemies.
If you don't add a projectile to the it, then it just won't shoot and will only look at the player.
It does so by getting the x and y difference in position then getting the angle using trig math.

The projectile angle thing is set up aswell as well as aligning the projectile with the the colour of the enemy (theoretically).
The projectile scripts rn aren't done, and there is nothing defining what colour is for enemies yet as far as I know.